+++ apps/demo/src/Views/Ui/components/TabView/utilities.ts
@@ -1,7 +1,10 @@
 export const addTab = (
   key: string,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   visibleTabs: any[],
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   setVisibleTabs: any,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   setActive: any,
 ) => {
   const existingTab = visibleTabs.find((tab) => tab === key);