+++ packages/ui/src/utils/UseExporter.ts
@@ -1,4 +1,4 @@
-import saveAs from "file-saver"; // eslint-disable-line import/no-named-as-default
+import saveAs from "file-saver";
 import { build, WorkSheetOptions } from "node-xlsx";
 import { useCallback, useState } from "react";
 
@@ -28,7 +28,7 @@ const exportXLSX = ({
   const buffer = build([{ name: sheetName, data, options: sheetOptions }]);
 
   saveAs(
-    new Blob([buffer], {
+    new Blob([buffer as BlobPart], {
       type: "application/vnd.ms-excel",
     }),
     `${filename}.xlsx`,